requirements.txt
opencv-python==4.5.5.64
face-recognition==1.3.0
dlib==19.24.0
streamlit==1.12.2
firebase-admin==6.1.0
google-cloud-storage==2.5.0
numpy==1.23.5
Pillow==9.3.0
python-dotenv==0.21.0
PyDrive==1.3.1